+++ src/roman_datamodels/testing/utils.py
@@ -389,7 +389,7 @@ def mk_gain(shape=None, filepath=None):
         return gainref
 
 
-def mk_mask(shape=True):
+def mk_mask(shape=None, filepath=None):
     meta = {}
     add_ref_common(meta)
     maskref = stnode.MaskRef()
@@ -401,7 +401,12 @@ def mk_mask(shape=True):
     else:
         maskref['dq'] = np.zeros((4096, 4096), dtype=np.uint32)
 
-    return maskref
+    if filepath:
+        af = asdf.AsdfFile()
+        af.tree = {'roman': maskref}
+        af.write_to(filepath)
+    else:
+        return maskref
 
 
 def mk_readnoise(shape=None, filepath=None):