+++ CHANGES.rst
@@ -3,6 +3,8 @@
 
 - Added readnoise maker utility and tests. [#23]
 
+- Added ReadNoise maker utility and tests. [#24]
+
 0.5.0 (2021-08-07)
 ==================