+++ CHANGES.rst
@@ -1,4 +1,4 @@
-0.6.0 (unreleased)
+0.5.1 (2021-08-24)
 ==================
 
 - Added tests for mask maker utility. [#25]
@@ -8,8 +8,6 @@
 - Added Readnoise maker utility and tests. [#23]
 
 - Added Gain maker utility and tests. [#24]
-
->>>>>>> 7b22bd94c43866c02f3d4662f2ce5679ca94d4fd
 
 0.5.0 (2021-08-07)
 ==================